refactor(tests): split content.test.ts into focused test modules
Replace the single 1127-line content.test.ts with four focused files:
content.core (performQualityAnalysis, message listener, getCssSelector,
getHtmlSnippet), content.accessibility, content.seo, and
content.performance. Each file imports only the functions it tests.
